- initial development of son-validate local api - changed directory structure: all api components are now located in 'src/son/validate/api'
